requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
numpy==1.21.4
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
toml==0.10.2
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
